+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.7.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Handle the situation of an empty DEM.
 
 
 1.7.0 (2023-01-11)