+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.7.1 (unreleased)
 ------------------
 
-- Handle the situation of an empty DEM.
+- Fixed breaches.line_id when there are boundary conditions in the model.
 
 
 1.7.0 (2023-01-11)