+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changelog of threedigrid-builder
 
 - Fixed breaches.line_id when there are boundary conditions in the model.
 
+- Revert "1D-2D lines derived from exchange lines are also converted to breaches"
+  from version 1.7.0.
+
 
 1.7.0 (2023-01-11)
 ------------------